fix: fixed validate methods to match the functionality of the command
